+++ app/components/components/Button/index.js
@@ -7,7 +7,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { theme } from '../../themes';
+import { theme } from '../../../themes';
 
 const ContainerButton = styled.button`
   width: ${props => props.width};
