+++ app/components/components/Header/index.js
@@ -7,8 +7,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { theme } from '../../themes';
-import search from '../../images/bus.png';
+import { theme } from '../../../themes';
+import search from '../../../images/bus.png';
 
 const ContainerHeader = styled.div`
   position: absolute;
