+++ app/components/components/MenuLateral/tests/index.test.js
