+++ app/components/components/Rectangulo/tests/index.test.js
