+++ app/components/components/Table/index.js
@@ -7,7 +7,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { theme } from '../../themes';
+import { theme } from '../../../themes';
 
 const ContainerTable = styled.div`
   width: 100%;
